requirements.txt
requests>=2.31.0
PyYAML>=6.0
pillow>=10.0.0
websocket-client>=1.6.0
google-generativeai>=0.3.2
